+++ ignition/include/gympp/gazebo/IgnitionEnvironment.h
@@ -61,7 +61,7 @@ public:
     bool render(RenderMode mode) override;
     std::optional<Observation> reset() override;
     std::optional<State> step(const Action& action) override;
-    std::vector<unsigned> seed(unsigned seed = 0) override;
+    std::vector<size_t> seed(size_t seed = 0) override;
 
     // Public APIs
     EnvironmentPtr env();
